Run the transaction with no test flags
RPMTRANS_FLAG_TEST is a double-edged sword.

We need this to test for file conflicts not picked up by zif,
but we also don't want to run the transaction twice.

Using RPMTRANS_FLAG_TEST also means that the pretrans scriptlets
are not run, which can cause the transaction to fail.