Add initial RHN package and store support
This functionality allows Zif to connect to RHN and for RHN to act
as a datastore in-place of a local metadata store.

The code is slow, and no requests are pipelined yet. It only accepts
a username and password as the authentication scheme, as the code is
in testing and machine registrations may be problematic at this
stage.

But it works. Weeeee....