Use wget to download files from FTP servers
FTP is a nasty protocol that libsoup doesn't support. For the few mirrors that
do not redirect to 'http://' use a wget helper to retrieve the file.

Doing this means we don't get any status completion, but there doesn't seem to
be any other credible FTP library other than libcurl. If somebody ported the
hacky wget code to use libcurl then that would be wonderful.