Remove the ZifDepend per-depend hash in ZifPackage to save the rainforests
It turns out that using the ZifDepend description as a hash key fails in a
couple of ways:

 * Doing the g_strdup_printf() for each depend in the hotpath isn't cool

 * Keeping a g_strdup'd version of the description increased memory requirements
   by about 30 bytes per ZifDepend. When you've got a few hundred thousand
   deps in a typical transaction, that eats up memory fast.

So, if we only hit the full depend hash one time in a hundred (just using the
name is 99% of all cache misses) then why do we bother?

Elad asked that question, and it turns out it was an excellent question to ask.

By not using the full ZifDepend cache for a transaction involving 130 packages
to update, the user time decreases from 39 seconds to 36. What's more dramatic
is the RSS memory use, down from a whopping 446Mb to a better 230Mb.

For that reason we want to avoid doing the secondary cache. Which is what this
commit does. :)