Try to make the signal handling in libzif somewhat sane
Basically, librpm steals SIGINT and a few other signals whenever it opens or
closes the rpmdb. To work around this, disconnect SIGINT using rpmsqEnable and
connect it to a cancel handler after any librpm transaction.

Whilst not ideal, this means that ctrl-c causes the GCancellable to be
cancelled at a sane point and allows zif to clean up and exit as soon as
possible. Using the GCancellable also means the GFile stuff is cancelled
correctly too.

Resolves https://bugzilla.redhat.com/show_bug.cgi?id=739814