Allow a transaction commit to install untrusted files
If we're explicitly showing the 'black flag' packages that are untrusted and ask
for confirmation, then it's okay to install packages without a key.

Add a new method zif_transaction_commit_full() that can be used by clients that
want control of the untrusted behaviour, which can be used in the PackageKit
backend.