Support per-thread and system-wide lock modes
A THREAD lock mode is used when another thread cannot access some shared state,
for instance when populating an in-memory cache.

A PROCESS lock mode is used when the package manager is doing something to the
system, which should be locked even for other processes.

This breaks API and ABI, again.