Speed up zif_store_array_get_updates() by a factor of 4.6
Typically, the size of the obsoletes table is small. By keeping a cached list of
the obsoletes that _do_ exist we can filter out depends that cannot exist when
using zif_store_what_obsoletes().

This limits the size of the SQL statement to a few hundred OR's per repo rather
than the thousands we were doing before.

This speeds up 'zif get-updates' substantially, on my system going from 9.45
seconds to 1.98 seconds.